requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
discord.py>=2.0.0
python-telegram-bot>=20.0
selenium>=4.0.0
apscheduler>=3.10.0
requests>=2.31.0 
webdriver-manager>=3.8.0
psutil>=5.9.0
supabase>=2.0.0
aiohttp>=3.8.0
PyYAML>=6.0.1
anyio>=3.6.2,<3.7.0
nest-asyncio>=1.5.0
flask==2.0.1